Remove datadir on node shutdown in DummyTask test, continued

+++ tests/golem/task/dummy/runner.py
@@ -64,7 +64,8 @@ def run_requesting_node(datadir, num_subtasks=3):
         client and client.quit()
         reactor.running and reactor.callFromThread(reactor.stop)
         logging.shutdown()
-        shutil.rmtree(datadir)
+        if os.path.exists(datadir):
+            shutil.rmtree(datadir)
 
     atexit.register(shutdown)
 
@@ -106,7 +107,8 @@ def run_computing_node(datadir, peer_address, fail_after=None):
         client and client.quit()
         reactor.running and reactor.callFromThread(reactor.stop)
         logging.shutdown()
-        shutil.rmtree(datadir)
+        if os.path.exists(datadir):
+            shutil.rmtree(datadir)
 
     atexit.register(shutdown)
